Improving adapt_template.cmake
- Add a DRY_RUN option
- Remove adapt_template.cmake script at the end.

+++ adapt_template.cmake
@@ -1,6 +1,8 @@
 # Check for NEW_NAME on command line (Required)
 if ("${NEW_NAME}" STREQUAL "")
-    message("Usage: cmake -DNEW_NAME=<new project name> -P adapt_template.cmake")
+    message("Usage: cmake -DNEW_NAME=<new project name>\n"
+            "             [-DDRY_RUN=1]\n"
+            "             -P adapt_template.cmake")
     message(FATAL_ERROR "NEW_NAME not provided")
 endif()
 
@@ -16,20 +18,38 @@ string(REPLACE "-" "_" NEW_NAME_UPPER "${NEW_NAME_UPPER}")
 
 # Validated and ready to go!
 message("Adapting template to new name: ${NEW_NAME}")
-find_package(Git REQUIRED)
+find_package(Git REQUIRED QUIET)
 
+# Helper function to rename a file
 function(rename_file OLD_PATH NEW_PATH)
     message("Rename ${OLD_PATH} -> ${NEW_PATH}")
+    if (${DRY_RUN})
+        return()
+    endif()
     execute_process(COMMAND "${GIT_EXECUTABLE}" mv "${CMAKE_CURRENT_LIST_DIR}/${OLD_PATH}" "${CMAKE_CURRENT_LIST_DIR}/${NEW_PATH}"
                     COMMAND_ERROR_IS_FATAL ANY)
 endfunction()
 
+# Helper function to replace occurrances of a string in a file
 function(replace_in_file FILE_PATH OLD_STRING NEW_STRING)
     message("Replace string in ${FILE_PATH}: \"${OLD_STRING}\" -> \"${NEW_STRING}\"")
+    if (${DRY_RUN})
+        return()
+    endif()
     file(READ "${CMAKE_CURRENT_LIST_DIR}/${FILE_PATH}" OLD_FILE_CONTENTS)
     string(REPLACE "${OLD_STRING}" "${NEW_STRING}" NEW_FILE_CONTENTS "${OLD_FILE_CONTENTS}")
     file(WRITE "${CMAKE_CURRENT_LIST_DIR}/${FILE_PATH}" "${NEW_FILE_CONTENTS}")
 endfunction()
+
+# Helper function to remove a file
+function(remove_file FILE_PATH)
+    message("Remove ${FILE_PATH}")
+    if (${DRY_RUN})
+        return()
+    endif()
+    execute_process(COMMAND "${GIT_EXECUTABLE}" rm "${CMAKE_CURRENT_LIST_DIR}/${FILE_PATH}"
+                    COMMAND_ERROR_IS_FATAL ANY)
+endfunction(remove_file)
 
 # Replace occurances of cpp_template in files
 replace_in_file(CMakeLists.txt cpp_template ${NEW_NAME})
@@ -45,3 +65,6 @@ replace_in_file(vcpkg.json cpp-template ${NEW_NAME_HYPENS})
 rename_file(include/cpp_template.h "include/${NEW_NAME}.h")
 rename_file(src/cpp_template.cpp "src/${NEW_NAME}.cpp")
 rename_file(test/cpp_template_test.cpp "test/${NEW_NAME}_test.cpp")
+
+# Remove this script
+remove_file(adapt_template.cmake)
